move some utility functions into a subpackage, for now import all
